webapi/tct-manual-w3c-tests/touchevent/TouchEvent_altKey_exist.html

<!DOCTYPE html>
<!--
Copyright (c) 2012 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Cui,Jieqiong <jieqiongx.cui@intel.com>

-->

<html>
  <head>
    <title>TouchEvent Test: TouchEvent_altKey_exist</title>
    <link rel="author" title="Intel" href="http://www.intel.com/" />
    <link rel="help" href="http://www.w3.org/TR/2013/WD-touch-events-20130124/#touchevent-interface" />
    <meta name="flags" content="interact" />
    <meta name="assert" content="Check if the attribute altKey exists" />
    <script src="../resources/blacklist.js" type="text/javascript"></script>
    <script src="../resources/testharness.js" type="text/javascript"></script>
    <script src="../resources/testharnessreport.js" type="text/javascript"></script>
  </head>
  <body onload="run()">
    <h1>Please touch the "Touch" button</h1>
    <button id="touchdiv">Touch</button>
    <div id="log"></div>
    <script>
        t = async_test(document.title, { timeout: 300000000000});
        function listener(ev) {
            t.step(function() {
                assert_true("altKey" in ev, "the attribute altkey exists");
            });
            t.done();
        }
        function run() {
            if (!is_platform_supported("TouchEventsManager.createTouch")) {
                try {
                    document.getElementById('touchdiv').addEventListener('touchstart', listener, false);
                } catch (err) {
                    t.step(function() {
                        if (err.name == "NOT_SUPPORTED_ERROR") {
                            assert_true(true, "Mark the case as \"PASS\" because the platform actually doesn't support the interface");
                        } else {
                            assert_true(false, "Exception while get altKey attribute" + err.name + err.message);
                        }
                    });
                    t.done();
                }
                return;
            }
            //The API is not in blacklist, then do normal tests
            try {
                document.getElementById('touchdiv').addEventListener('touchstart', listener, false);
            } catch (err) {
                t.step(function() {
                    assert_true(false, "Exception while get altkey attribute " + err.name + err.message);
                });
                t.done();
            }
        }
    </script>
  </body>
</html>
